[GREEN] Add VerifyForm.tsx and change email verification page.
